Added feature to show todo item from user defined projects
Created another array in the todo.js file to store todo items which have
been assigned with a project index. Then this array will store array of
todo items based on the projectIndex property of the todo item.

For example, a todo item with projectIndex of 2 with be assigned to the
subarray of index 2 inside the main array.