Refactor code to enable more dynamic manipulation over todo elements and their DOM
A feature is added to update the current dashboard by adding a flag to
track what the active project is before creating a new todo item.

However, the way the code was previously structured makes it hard to
track, retrieve and change what the active project. Therefore, the data
structure has been modified for todo items which is expected to
faciliate better data storage and retrieval at a later stage also.