Refactor code to facilitate the edit todo feature
The previous code is too bulky because a dialog and form is added for
each form that requires user's input e.g. add todo, add project, edit
todo. The code now includes just one dialog where a form will only be
added to the dialog when it's needed.

This way a new form will be created and added to the dialog only when
the edit todo button is clicked. And the code can then read the value of
this specific form and update the existing todo item accordingly.