Fixed bug where todo items do not show in their associated project correctly
When a project is deleted, a function will update the projectIndex
property of those todo items that are of an index larger than the
activeProject project index. So that the todos before the deleted
project will not be modified but the ones after will have its
projectIndex deducted by one to close up the array.