Added module to store and access data in localStorage
Note that the the todo module was instantiated in the handler module
which most of the functions in the handler module depends on the data
stored in that instance. For example, the array where all todo items is
stored. Therefore, methods are created in project and todo modules so
they can be used in the handler module to modify the data stored in
those instances.

By checking whether localStorage has relevant data stored, data from
localStorage will be then be used to modify the data of the project and
todo module instances. Hence the data can be retrieved and used again
after page refresh.